fix getGuard

+++ nodes/record_mgr.h
@@ -69,7 +69,7 @@ public:
     RecordMgr(const RecordMgr&) = delete;
 
     auto getGuard() const {
-        return 0;
+        return size_t(0);
     }
 
     LNodeWrapper<key_t, val_t> get_new_node(key_t key) const {
